Renamed serviceStore.test to service.test and adapts to these changes
